🧩 :: (#581) company 권한 검사 추가

+++ jobis-application/src/main/java/team/retum/jobis/domain/recruitment/usecase/DeleteRecruitAreaUseCase.java
@@ -2,11 +2,14 @@ package team.retum.jobis.domain.recruitment.usecase;
 
 import lombok.RequiredArgsConstructor;
 import team.retum.jobis.common.annotation.UseCase;
+import team.retum.jobis.common.spi.SecurityPort;
 import team.retum.jobis.domain.recruitment.exception.RecruitAreaNotFoundException;
 import team.retum.jobis.domain.recruitment.model.RecruitArea;
 import team.retum.jobis.domain.recruitment.service.CheckRecruitmentAreaDeletableService;
+import team.retum.jobis.domain.recruitment.service.CheckRecruitmentPermissionService;
 import team.retum.jobis.domain.recruitment.spi.CommandRecruitmentPort;
 import team.retum.jobis.domain.recruitment.spi.QueryRecruitmentPort;
+import team.retum.jobis.domain.user.spi.QueryUserPort;
 
 @RequiredArgsConstructor
 @UseCase
@@ -14,11 +17,13 @@ public class DeleteRecruitAreaUseCase {
     private final CommandRecruitmentPort commandRecruitmentPort;
     private final QueryRecruitmentPort queryRecruitmentPort;
     private final CheckRecruitmentAreaDeletableService checkRecruitmentAreaDeletableService;
+    private final CheckRecruitmentPermissionService checkRecruitmentPermissionService;
 
     public void execute(Long recruitAreaId) {
         RecruitArea recruitArea = queryRecruitmentPort.queryRecruitmentAreaById(recruitAreaId)
                 .orElseThrow(() -> RecruitAreaNotFoundException.EXCEPTION);
         checkRecruitmentAreaDeletableService.checkRecruitmentAreaDeletable(recruitArea);
+        checkRecruitmentPermissionService.checkPermission(recruitArea);
 
         commandRecruitmentPort.deleteRecruitAreaById(recruitArea.getId());
     }
