+++ jobis-application/src/main/java/team/retum/jobis/domain/recruitment/usecase/DeleteRecruitmentUseCase.java
@@ -2,8 +2,10 @@ package team.retum.jobis.domain.recruitment.usecase;
 
 import lombok.RequiredArgsConstructor;
 import team.retum.jobis.common.annotation.UseCase;
+import team.retum.jobis.common.spi.SecurityPort;
 import team.retum.jobis.domain.recruitment.exception.RecruitmentNotFoundException;
 import team.retum.jobis.domain.recruitment.model.Recruitment;
+import team.retum.jobis.domain.recruitment.service.CheckRecruitmentPermissionService;
 import team.retum.jobis.domain.recruitment.spi.CommandRecruitmentPort;
 import team.retum.jobis.domain.recruitment.spi.QueryRecruitmentPort;
 
@@ -12,11 +14,13 @@ import team.retum.jobis.domain.recruitment.spi.QueryRecruitmentPort;
 public class DeleteRecruitmentUseCase {
     private final QueryRecruitmentPort queryRecruitmentPort;
     private final CommandRecruitmentPort commandRecruitmentPort;
+    private final CheckRecruitmentPermissionService checkRecruitmentPermissionService;
 
     public void execute(Long recruitmentId) {
         Recruitment recruitment = queryRecruitmentPort.queryRecruitmentById(recruitmentId)
                 .orElseThrow(() -> RecruitmentNotFoundException.EXCEPTION);
         recruitment.checkIsDeletable();
+        checkRecruitmentPermissionService.checkPermission(recruitment);
 
         commandRecruitmentPort.deleteRecruitment(recruitment);
     }
